requirements.txt
flask==3.1.2
pandas==2.3.2
numpy==2.3.2
scikit-learn==1.8.0
matplotlib==3.10.6
plotly==6.5.0
firebase-admin==7.1.0
python-dotenv==1.2.1
flask-cors==6.0.2
